# public/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Hifive football api",
    "version": "1.0.0",
    "title": "HifiveFootball",
    "termsOfService": "http://api.hifivefootball.com/",
    "contact": {
      "email": "ddastudio@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "api.hifivefootball.com:5300",
  "basePath": "/v1",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/auth/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "content",
        "description": "login",
        "operationId": "login by id, pw",
        "consumes": [
          "application/x-www-form-urlencoded"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "formData",
            "name": "username",
            "description": "userid",
            "required": true,
            "default": "2222",
            "type": "string"
          },
          {
            "in": "formData",
            "name": "password",
            "description": "pw",
            "required": true,
            "type": "string",
            "default": "2222"
          }
        ],
        "responses": {
          "200": {
            "description": "login success",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/AuthModel"
              }
            }
          },
          "400": {
            "description": "login fail",
            "schema": {
              "type": "object",
              "items": {
                "$ref": "#/definitions/ApiResponse2"
              }
            }
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "petstore_auth": {
      "type": "oauth2",
      "authorizationUrl": "http://petstore.swagger.io/oauth/dialog",
      "flow": "implicit",
      "scopes": {
        "write:pets": "modify pets in your account",
        "read:pets": "read your pets"
      }
    },
    "api_key": {
      "type": "apiKey",
      "name": "api_key",
      "in": "header"
    }
  },
  "definitions": {
...
    "ApiResponse2": {
      "type": "object",
      "properties": {
        "result": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "AuthModel": {
      "type": "object",
      "properties": {
        "access_token": {
          "type": "string"
        },
        "refresh_token": {
          "type": "string"
        },
        "expires_in": {
          "type": "integer",
          "format": "int32"
        },
        "token_type": {
          "type": "string"
        }
      }
    },
...
  },
...
}